Fix object storage benchmark tests
Move all tests to the correct file and update them for the recent
refactor.
